backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
sentence-transformers
faiss-cpu
numpy